# conf/openapi.yaml
openapi: "3.0.3"
info:
  title: Deposplit API
  version: "0.1"
  description: |
    REST API for the Deposplit secret-sharing Web app/service.

    ## Authentication

    Every request must carry three headers:

    | Header | Content |
    |---|---|
    | `X-Deposplit-Verify-Key` | Caller's public verify key (base64url — see `PublicKey` schema) |
    | `X-Deposplit-Nonce` | Per-request unique string in the form `<unix-ms>.<random>` — server rejects requests whose embedded timestamp is more than 5 minutes old |
    | `X-Deposplit-Signature` | Base64url Ed25519 signature (RFC 8032) |

    The signature covers the canonical string:

    ```
    nonce || "\n" || UPPERCASE(method) || "\n" || path_with_query || "\n" || hex(SHA-256(body))
    ```

    where `body` is the empty string for requests without a body.
    The Web app/service verifies the signature against the verify key supplied in `X-Deposplit-Verify-Key`.
    No pre-registration is required — the keypair is the identity.

    ## Payload signatures

    Two additional Ed25519 signatures ride with every `ShareRequest` row, independent of the
    transport-auth signature above. The transport signature authenticates the HTTP caller for one
    specific call and is never persisted — it gives a later reader of a row (via `GET`) nothing to
    re-verify authorship against. `senderSignature` and `recipientSignature` close that gap: any
...
    through a relay other than deposplit.com, e.g. a self-hosted instance) safe — a passive
...
servers:
...
security:
  - ed25519Auth: []
components:
  securitySchemes:
    ed25519Auth:
      type: apiKey
      in: header
      name: X-Deposplit-Signature
      description: >
        Must accompany X-Deposplit-Verify-Key and X-Deposplit-Nonce headers.
        See top-level description for the canonical signing string.
  schemas:
...
    PublicKey:
      type: string
      description: >
        Base64url-encoded signing-algorithm public key (32 bytes, no padding, for the current
        cipher suite `ed25519+x25519-v1` — see `CipherSuite`) — used for API authentication and as
        the routing identity. Not pinned to an exact length; a future cipher suite may use a
        different key size.
    CipherSuite:
      type: string
      enum: [ed25519+x25519-v1]
      description: >
        The matched pairing of signing algorithm + key-agreement algorithm an identity currently
        uses. One value exists
        today; the type exists so a future fleet-wide algorithm swap is an additive new value
        rather than a breaking wire-format migration.
...
    X25519PublicKey:
      type: string
      description: >
        Base64url-encoded key-agreement public key (32 bytes, no padding, for the current cipher
        suite `ed25519+x25519-v1` — see `CipherSuite`), used for share encryption client-side. The
        Web app/service never performs key agreement with it — it only ever routes it opaquely,
        same as ciphertext. Not pinned to an exact length; a future cipher suite may use a
        different key size.
    KeyRotation:
      description: >
        The signed `rotate(K_old -> K_new)` push — a holder's proactive "I am now
        newVerifyKey, previously oldVerifyKey" notice, addressed to one contact
        (`recipientKey`) at a time. Deliberately not a `ShareRequest`: it carries no `secretId`
        and has no consent phase — the recipient auto-verifies `signature` against
        `oldVerifyKey` (the trusted key it already knows this contact by) and, on success,
        updates its local contact record in place before deleting the row.
      type: object
      required: [id, oldVerifyKey, recipientKey, newVerifyKey, newEncKey, newCipherSuite, signature, createdAt]
      properties:
        id:
          type: string
          format: uuid
        oldVerifyKey:
          $ref: '#/components/schemas/PublicKey'
        recipientKey:
          $ref: '#/components/schemas/PublicKey'
        newVerifyKey:
          $ref: '#/components/schemas/PublicKey'
        newEncKey:
          $ref: '#/components/schemas/X25519PublicKey'
        newCipherSuite:
          $ref: '#/components/schemas/CipherSuite'
        signature:
          type: string
          description: >
            Base64url signature by oldVerifyKey — see "Payload signatures" above.
        createdAt:
          type: string
          format: date-time
    PushRotationBody:
      type: object
      required: [recipientKey, newVerifyKey, newEncKey, newCipherSuite, signature]
      properties:
        recipientKey:
          $ref: '#/components/schemas/PublicKey'
        newVerifyKey:
          $ref: '#/components/schemas/PublicKey'
        newEncKey:
          $ref: '#/components/schemas/X25519PublicKey'
        newCipherSuite:
          $ref: '#/components/schemas/CipherSuite'
        signature:
          type: string
          description: >
            Base64url signature by the caller over this notice's canonical bytes — see
            "Payload signatures" above. Returns `400` if it doesn't verify.
    CustodyHeartbeat:
      description: >
        The signed custodial-heartbeat push — a holder's proactive "still guarding
        {secretIds} for you" notice (or, when optedOut is true, a signed "my silence from here on
        is not a loss signal" notice), addressed to one owner at a time. Deliberately not a
        `ShareRequest`/`KeyRotation`: no singular `secretId`, no consent phase, and — unlike
        `KeyRotation`/inventory — never consumed-and-deleted; the Web app/service keeps only the
        *latest* heartbeat per `(holderKey, ownerKey)` pair, upserted on every push.
      type: object
      required: [id, holderKey, ownerKey, secretIds, optedOut, signature, createdAt]
      properties:
        id:
          type: string
          format: uuid
        holderKey:
          $ref: '#/components/schemas/PublicKey'
        ownerKey:
          $ref: '#/components/schemas/PublicKey'
        secretIds:
          type: array
          items:
            type: string
            format: uuid
          description: Every secret this holder currently holds a share of for this owner.
        optedOut:
          type: boolean
          description: >
            When true, this holder has signalled it will no longer heartbeat this owner — the
            owner's app should stop treating subsequent silence from this holder as a loss signal.
        signature:
          type: string
          description: >
            Base64url Ed25519 signature by holderKey — see "Payload signatures" above.
        createdAt:
          type: string
          format: date-time
    PushHeartbeatBody:
      type: object
      required: [ownerKey, secretIds, optedOut, signature]
      properties:
        ownerKey:
          $ref: '#/components/schemas/PublicKey'
        secretIds:
          type: array
          items:
            type: string
            format: uuid
        optedOut:
          type: boolean
        signature:
          type: string
          description: >
            Base64url Ed25519 signature by the caller over this heartbeat's canonical bytes — see
            "Payload signatures" above. Returns `400` if it doesn't verify.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
paths:
...
  /key-rotations:
    post:
      operationId: pushRotation
      summary: Push a signed key-rotation notice to one contact
      description: >
        The authenticated caller becomes `oldVerifyKey`. Not consent-gated and no conflict check
        — a fire-and-forget push, the same shape as `inventory`: the recipient
        polls, auto-verifies, and deletes the row once consumed.
      tags: [KeyRotations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushRotationBody'
      responses:
        '201':
          description: Notice pushed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyRotation'
        '400':
          description: Missing/invalid field, or signature missing/invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listRotations
      summary: List rotation notices addressed to the caller
      tags: [KeyRotations]
      responses:
        '200':
          description: Rotation notice list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KeyRotation'
  /key-rotations/{id}:
    delete:
      operationId: deleteRotation
      summary: Delete a rotation notice once consumed
      description: >
        Only the recipient may delete it — unlike `ShareRequest` rows, there's no sender-side
        reason to read one back.
      tags: [KeyRotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Notice deleted.
        '403':
          description: Caller is not the recipient of this notice.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Notice not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /custody-heartbeats:
    post:
      operationId: pushHeartbeat
      summary: Push (upsert) a signed custodial-heartbeat for one owner
      description: >
        The authenticated caller becomes `holderKey`. Not consent-gated and no conflict check —
        a fire-and-forget push, the same shape as `inventory`/`key-rotations`. Unlike those,
        never consumed-and-deleted: this upserts by `(holderKey, ownerKey)`, replacing any
        previous heartbeat from the same holder to the same owner rather than accumulating.
      tags: [CustodyHeartbeats]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushHeartbeatBody'
      responses:
        '201':
          description: Heartbeat pushed (created or updated).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustodyHeartbeat'
        '400':
          description: Missing/invalid field, or signature missing/invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listHeartbeats
      summary: List the latest heartbeat from each holder addressed to the caller
      tags: [CustodyHeartbeats]
      responses:
        '200':
          description: Heartbeat list, one entry per distinct holder.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustodyHeartbeat'